zephyr: no need for explicit memory sharing
When built with Zephyr all memory except stack is sharable: either it
is read-only, or it is accessed via the non-caching address range.
Besides no .shared_data section is actually created under Zephyr,
therefore trying to place data in it leads to linker warnings and
actually moves that data to the caching address range.

Signed-off-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>